updated icons to .svgz, fix crash, update formats
There was an issue upon dragging widgets to the plasma desktop, which was caused by the padding in the dialog/background.svgz file. This change does slightly change the look of the theme, but this was a necessary change to fix said issue. But i will say, it does looks nice! :smile: